Header search and autosuggest file and structure cleanup, redux-first-router actions removed and split into util functions with tests
