+++ src/app/components/DataCard/DataCard.tsx
@@ -13,7 +13,7 @@ import type { FunctionComponent } from 'react'
 import { Fragment } from 'react'
 import { Link as RouterLink } from 'react-router-dom'
 import styled from 'styled-components'
-import { getDetailPageData } from '../../../store/redux-first-router/actions'
+import getDetailPageData from '../../utils/getDetailPageData'
 import { toDataDetail, toDataSearch } from '../../links'
 import type { DataResult } from '../../pages/SearchPage/types'
 import formatCount from '../../utils/formatCount'
