+++ src/app/components/DataCard/DataList.tsx
@@ -2,7 +2,7 @@ import { breakpoint, Link, themeSpacing } from '@amsterdam/asc-ui'
 import type { FunctionComponent } from 'react'
 import { Link as RouterLink } from 'react-router-dom'
 import styled from 'styled-components'
-import { getDetailPageData } from '../../../store/redux-first-router/actions'
+import getDetailPageData from '../../utils/getDetailPageData'
 import { toDataDetail, toDataSearch } from '../../links'
 import { activeFiltersParam } from '../../pages/SearchPage/query-params'
 import type { DataResult } from '../../pages/SearchPage/types'
