+++ src/app/components/DataSelection/DataSelectionList.tsx
@@ -2,7 +2,7 @@ import { Link, themeSpacing } from '@amsterdam/asc-ui'
 import { Link as RouterLink } from 'react-router-dom'
 import styled from 'styled-components'
 import type { FunctionComponent } from 'react'
-import { getDetailPageData } from '../../../store/redux-first-router/actions'
+import getDetailPageData from '../../utils/getDetailPageData'
 import buildDetailUrl from '../../pages/MapPage/components/DetailPanel/buildDetailUrl'
 import DataSelectionFormatter from './DataSelectionFormatter/DataSelectionFormatter'
 import type { Data } from './types'
