+++ src/app/components/Header/Header.tsx
@@ -7,11 +7,9 @@ import {
 } from '@amsterdam/asc-ui'
 import styled, { css } from 'styled-components'
 import type { FunctionComponent } from 'react'
-import HeaderSearch from '../../../header/components/HeaderSearch'
+import { HeaderSearchProvider } from './HeaderSearchProvider'
+import HeaderSearch from './HeaderSearch'
 import HeaderMenu from './HeaderMenu'
-
-// TODO relative to this src/app/components/header/Header.tsx this dir structure is rather confusing
-import { HeaderSearchProvider } from '../../../header/components/HeaderSearchProvider'
 
 const stickyStyle = css`
   position: sticky;
