+++ src/app/components/Header/HeaderSearch.tsx
@@ -4,17 +4,16 @@ import { useCallback, useRef, useState } from 'react'
 import { useHistory } from 'react-router-dom'
 import styled from 'styled-components'
 import debounce from 'lodash.debounce'
-import SearchBar from '../../app/components/SearchBar'
-import { LOCAL_STORAGE_KEY } from '../../app/components/SearchBarFilter/SearchBarFilter'
-import SEARCH_PAGE_CONFIG from '../../app/pages/SearchPage/config'
-import { SearchType } from '../../app/pages/SearchPage/constants'
-import { queryParam } from '../../app/pages/SearchPage/query-params'
-import toSearchParams from '../../app/utils/toSearchParams'
-// import useParam from '../../app/utils/useParam'
+import SearchBar from '../SearchBar'
+import { LOCAL_STORAGE_KEY } from '../SearchBarFilter/SearchBarFilter'
+import SEARCH_PAGE_CONFIG from '../../pages/SearchPage/config'
+import { SearchType } from '../../pages/SearchPage/constants'
+import { queryParam } from '../../pages/SearchPage/query-params'
+import toSearchParams from '../../utils/toSearchParams'
 import { useHeaderSearch } from './HeaderSearchProvider'
-import useTraverseList from '../../app/utils/useTraverseList'
-import type { AutoSuggestSearchResult } from '../services/auto-suggest/auto-suggest'
-import autoSuggestSearch, { MIN_QUERY_LENGTH } from '../services/auto-suggest/auto-suggest'
+import useTraverseList from '../../utils/useTraverseList'
+import type { AutoSuggestSearchResult } from './services/auto-suggest/auto-suggest'
+import autoSuggestSearch, { MIN_QUERY_LENGTH } from './services/auto-suggest/auto-suggest'
 import AutoSuggest from './auto-suggest/AutoSuggest'
 
 const StyledLegend = styled.legend`
