+++ src/app/components/Header/auto-suggest/AutoSuggest.tsx
@@ -2,14 +2,14 @@ import { Button, Paragraph, themeSpacing } from '@amsterdam/asc-ui'
 import { forwardRef, useMemo } from 'react'
 import { useHistory } from 'react-router-dom'
 import styled from 'styled-components'
-import LoadingSpinner from '../../../app/components/LoadingSpinner/LoadingSpinner'
-import { toSearch } from '../../../app/links'
-import { SearchType } from '../../../app/pages/SearchPage/constants'
-import { queryParam } from '../../../app/pages/SearchPage/query-params'
-import toSearchParams from '../../../app/utils/toSearchParams'
-import type { CmsType } from '../../../shared/config/cms.config'
-import type { AutoSuggestSearchResult } from '../../services/auto-suggest/auto-suggest'
-import { LABELS } from '../../services/auto-suggest/auto-suggest'
+import LoadingSpinner from '../../LoadingSpinner/LoadingSpinner'
+import { toSearch } from '../../../links'
+import { SearchType } from '../../../pages/SearchPage/constants'
+import { queryParam } from '../../../pages/SearchPage/query-params'
+import toSearchParams from '../../../utils/toSearchParams'
+import type { CmsType } from '../../../../shared/config/cms.config'
+import type { AutoSuggestSearchResult } from '../services/auto-suggest/auto-suggest'
+import { LABELS } from '../services/auto-suggest/auto-suggest'
 import AutoSuggestCategory from './AutoSuggestCategory'
 
 export type SearchCategory = CmsType | SearchType
