+++ src/app/components/Header/auto-suggest/AutoSuggestCategory.tsx
@@ -1,14 +1,10 @@
 import { Link } from 'react-router-dom'
 import type { LocationDescriptorObject } from 'history'
 import type { FunctionComponent } from 'react'
-import SEARCH_PAGE_CONFIG from '../../../app/pages/SearchPage/config'
-import {
-  activeFiltersParam,
-  pageParam,
-  queryParam,
-} from '../../../app/pages/SearchPage/query-params'
-import toSearchParams from '../../../app/utils/toSearchParams'
-import type { AutoSuggestSearchResult } from '../../services/auto-suggest/auto-suggest'
+import SEARCH_PAGE_CONFIG from '../../../pages/SearchPage/config'
+import { activeFiltersParam, pageParam, queryParam } from '../../../pages/SearchPage/query-params'
+import toSearchParams from '../../../utils/toSearchParams'
+import type { AutoSuggestSearchResult } from '../services/auto-suggest/auto-suggest'
 import AutoSuggestItem from './AutoSuggestItem'
 
 export interface AutoSuggestCategoryProps {
