+++ src/app/pages/MapPage/components/MapSearchResults/MapSearchResults.tsx
@@ -5,7 +5,7 @@ import { Link as RouterLink } from 'react-router-dom'
 import styled from 'styled-components'
 import { getScopes } from '../../../../../shared/services/auth/auth'
 import formatNumber from '../../../../../shared/services/number-formatter/number-formatter'
-import { getDetailPageData } from '../../../../../store/redux-first-router/actions'
+import getDetailPageData from '../../../../utils/getDetailPageData'
 import AuthAlert from '../../../../components/Alerts/AuthAlert'
 import LoadingSpinner from '../../../../components/LoadingSpinner/LoadingSpinner'
 import ShowMore from '../../../../components/ShowMore'
