+++ src/app/pages/MapPage/legacy/services/map-services.config.ts
@@ -20,7 +20,7 @@ import { fetchWithToken } from '../../../../../shared/services/api/api'
 import AuthScope from '../../../../../shared/services/api/authScope'
 import type { Wsg84Coordinate } from '../../../../../shared/services/coordinate-reference-system/crs-converter'
 import getRdAndWgs84Coordinates from '../../../../../shared/services/coordinate-reference-system/getRdAndWgs84Coordinates'
-import { getDetailPageData } from '../../../../../store/redux-first-router/actions'
+import getDetailPageData from '../../../../utils/getDetailPageData'
 import formatDate from '../../../../utils/formatDate'
 import getFileName from '../../../../utils/getFileName'
 import toSearchParams from '../../../../utils/toSearchParams'
